requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==23.0.0
python-dotenv==1.0.0
pydantic==1.10.12  # Using older version that doesn't require Rust
requests==2.31.0
twilio==8.10.0

# SQLite instead of PostgreSQL for now
sqlalchemy==2.0.23
# No asyncpg or psycopg2

# MongoDB
pymongo==4.6.0
motor==3.3.1

# Authentication - simplified
python-jose==3.3.0
passlib==1.7.4

# Utility
email-validator==2.1.0
python-multipart==0.0.6
